Convirting TransferFormat in favor of TransferFrormatInto and TransferFormatTryFrom
TransferFormat previously cause multiple allocations which were generally not nessicary
besides being easier to implement. The new way allows for only one buffer to be created
which is then filled.

TransferFormat has also been split into *Into and *TryFrom so that things that can
implement converting from the transfer format or into the transfer format.

The server is now made of ServerAttribute trait objects so the user can perform some
operations before the data is written or read from the server. Most of the time this
will be locking and reference counting so that attributes can be used in multiple
servers for multiple connections.