Fixed code to pass tests.
Changes made to reading and writing attributes broke the
read/write reqeust processing in the attribute server. Now
the write will send back a response, and the read required
the the functions to be implemented in the implementation of
the trait TransferFormatInto for &dyn TransferFormatInto.

The permissions are *probably* also fixed, by the testing
isn't the best. It doesn't fully test all possible permission
permutations due to the runtime of the test on one thread being
atrocious.