Trying a weak matcher to try and clean out any saved matcher info that no longer has a future awaiting.
This branch is attempts to fix a bug where a matcher becomes
forgotten due to a future never polling to completion. The prior
implementation would only remove matchers from an associated
event (or None for all events) when the matcher data was taken
by a re-call to expect_event. If that re-call never happens
then the matchers would stay and cause problems when 'zombie'
matchers would be matched over currently in used matchers by the
event handler.