Moving the advertising modules around and added OOB data.
The moving the advertising modules around was decided as having all the AD types as modules in a single file is cumbersome. Added the AD (or EIR data struct, or whatever they're really called) for LE secure connection OOB data. Also added oob_block for creating the oob_block that contains these EIR data when sent as part of pairing.