Added the ability to externally provide the received OOB data to the security managers.
Before the security managers would just take a method that would create a future to await the reception of OOB data. Unfortunately this cannot be used in all cases with the security manager as at times the user may not be able to provide this method (due to things like lifetimes, mutablility, ect.). So the external method was added for the user to set the OOB data external to the normal operation of both security managers.