Implementing serde
This implementation of a deserializer is similar to the crate postcard except it adds hints to length meta information. Serializable types can skip serializing lengths if the length can be inferred in a different way. Often times the length in a header can be used to infer the length of a string, array, sequence, or map.