Adding interface module
This module is the start of the change to be
made of how the interface interacts with the
host. The 'interface' will soon be in its own
async task separate from where HCI commands
are issued. A 'host' async task will be where commands and events are processed by the
user. 'connection' async tasks will mostly
remain the same (they should still look the
same to the library user).