Updating the HCI command implementation for the changes to HostInterface.
The HostInterface will no longer have flow control or a
single send command.

Flow control will be part of the interface async task for
HCI commands or part of connection async tasks for HCI
data packets.

Host interface now two types of send commands. One for
commands that are responded with a command complete
event, and the other is for commands that are
responded with a command status event.