Implementing channel ends
The async tasks are being implemented to have bi-directional communication between the interface async task and one of the other async tasks. The was done with the interface having a single receiver, but that doesn't work when working with flow control in the controller. The new scheme is to block receiving from async tasks until the interface async task knows that the controller can accept a message.