Bulk of local_channel implementation is completed
This (mostly) completes the implementation for ChannelReserve for both local_dynamic_channel and local_stack_channel. There may be more implementation that needs to be done as the interface is further flushed out.

# Note
Still need to clean up warnings, but they will probably be cleaned up after receiving is done.