Removing `get_mut_reserve` in favor of having a trait to just init channel ends for the host task.
`get_mut_reserve` would need to be public if module interface is to become its own crate. Giving a mutable reference to the channel_reserve is too much access, as `get_mut_reserve` was just used to create the channel ends for the host async task.

The trait `InitHostTaskEnds` was created to replace `get_mut_reserve`, and it's only job is to create the channel ends for the host async task.