limiting the commands to one
For now the maximum number of commands that can be sent to the Controller is limited to one. Right now the disregarding of command response events (Command Complete / Command Status) is a very simplistic boolean (see field bo_tie_hci_interface::Interface::skip_cmd_response). A more complex processing of what command response events are dropped or sent to host is needed before multiple commands can be sent to the host.