Adding and reworking authentication
This adds the authentication methods number comparison
and passkey and reworks out of band. The previous way
was to have authentication integrated into the security
manager, but this can cause deadlocking in the connection
channel awaiting user or out of band input. The async
part of authentication is now outside the scope of the
security manager and pseudo-callbacks are used instead.

no changes have been tested yet.