Adding errors for intervals
These errors will eventually be the errors returned by the methods try_from_raw and try_from_duration for an interval but for now they are just converted into a string. Once Strings are no longer used as part of the error returned when trying to convert raw hci data into an event type, they will be set as the errors for these methods.