Adding "host" privacy to le-bonding-peripheral
"host" privacy is where the host generates and validates resolvable private addresses. When the controller does it, it uses a resolving list and this was previously implemented. To target Controllers that do not support the LE feature LL Privacy a basic host implementation of the resolving list was added to the example.