fixing case where disconnect is sent to connection after it is closed.
It turns out the controller can send the event 'connection terminated by local host' to the host. IDK why this is done, but it is something available to controllers. Also fixed disconnect being sent again to the connection async task if routing is enabled (which probably was causing a case were the sender (or receiver if the message was unwrapped) to panic).