changing the `Write` type of `AccessValue` to be a future that outputs a result.
It is within specification for a protocol above the ATT protocol to be able to return custom errors. These errors really only occur whenever the client writes the wrong value to the server.

In the future the type `Read` may also return an error but for now this is not implemented unless there is found a need to do so.