fixing Send safety
The autogenerated future would capture a `*mut P` across the await. The compiler would then complain that `*mut T` was not Send safe. This change causes the generated future to not capture the `*mut P` and instead 'captures' a `&mut T` which the compiler can currently handle for Send checks.