changed the output of the future returned by `AccessValue::read` to Result<_, crate::pdu::Error>
This allows for cases where the Client is sent an application or common error from a read request. From the server side, this isn't simple. A new trait `ReadGuard` is required do ensure that the implementation knows when the value is *read* for Server operational checks and when the value is *read* for transferring the data to the Client.

Also fixed an example.