Changing how credit based channel reception is done
Now the future returned by the method `receive` will maybe output the data. If input check_credits is `false` it will only output the `sdu`, but if it is `true` then `None` will be returned if the credit channel detects that the peer device is out of credits for sending on this channel.

Also fixed the `IntoIterator` and `ExactSizeIterator` implementations for `FlowControlCreditIndIter`