Fixing the LeLogicalLinkHandle's implementation of `LogicalLinkPrivate::new_dyn_channel`.
Before it assumed that the list was fully sized with `Unused`, and this was briefly the case when developing the logic, but now the internal channel list grows as channels are added.

New implementation first checks if the list is fully sized before searching all possible entries for a possible new channel.