Fixing how SDUs were recombined.
The bug was essentially the same buffer for the PDU was used for the SDU. It was trying to extend the buffer with itself essentially. This splits the PDU and SDU buffers. Now there is only one PDU buffer (only one PDU per link can be received at a time), but there is a SDU buffer for each SDU channel (for now that just means for each credit based channel).

CreditBasedChannelData::process_pdu also needed to be fixed as it was not calculating the correct number of bytes to take.