Fixing a bug with the MPS with credit based channels.
The channel was using the physical layer's MTU as the MPS when it should use the value defined as part of the L2CAP connection establishment handshake. The fragmentation size of the `PhysicalLayer` is not related to the MPS unless forced to be by the user's use of the library.

The consequently also meant that `send_pdu` does not require the `fragmentation_size` input. Only the bug used the field for another value other than the `PhysicalLink`'s maximum transmission size.