fixing method `next` to not be stateful
`stateful` here means there was data within the future returned by `next` that was important to link related state. The issue with this is fairly easily seen with select:

select! {
    _ = link.next() => /* do stuff */
    _ = other_future => {
        /* oh no my next states were dropped* /
    }
}

Having any state within the future returned by next is a problem as if it gets dropped then the processing of fragments gets messed up. The fix is to move all the states into the logical link.