fixing HashValue
Bug was that the return of `get_value` (@266 -> @266:271) was being treated as the transfer format value. The return is an Option<T> so this was incorrectly adding a pre-byte for the option. The fix is to use an unwarp.

Adding tests for checking the generated has value. These tests necessitated the creation of the method `create_glob`, which is really just what was done in `generate`.